+++ src/MyMediaLite/IO/MovieLensRatingData.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2010, 2011 Zeno Gantner
+// Copyright (C) 2010, 2011, 2012 Zeno Gantner
 //
 // This file is part of MyMediaLite.
 //
@@ -69,7 +69,7 @@ namespace MyMediaLite.IO
 
 				int user_id = user_mapping.ToInternalID(long.Parse(tokens[0]));
 				int item_id = item_mapping.ToInternalID(long.Parse(tokens[1]));
-				double rating = double.Parse(tokens[2], CultureInfo.InvariantCulture);
+				float rating = float.Parse(tokens[2], CultureInfo.InvariantCulture);
 				long seconds = uint.Parse(tokens[3]);
 
 				var time = new DateTime(seconds * 10000000L).AddYears(1969);
